% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigo-params.R
\name{codigo-params}
\alias{codigo-params}
\title{Common parameters used across codigo functions}
\arguments{
\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
Default is \code{v2}.}

\item{release}{A string specifying the release version to search from. If not
specified, defaults to the latest available release version. See the
available versions for each ICD classification with \code{icd_versions}.}

\item{linearization}{A character value for which linearization to search.
Currently, the possible values are \emph{"mms"} and \emph{"icf"}. Default is \emph{"mms"}.}

\item{language}{ICD-API is multi-lingual. Language codes such as "en", "es",
"zh", etc. Default is English ("en").}

\item{tabular}{Logical. Should output be structured into a tibble? Default
is TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be printed?
Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int.}

\item{client}{The OAuth2 client produced through a call to
\code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{timeout}{The maximum time to wait for a response from the API server in
seconds. Default is 30 seconds.}

\item{max_tries}{Maximum number of times to retry a request if it fails with
a transient error (e.g. 429, 500, 503). Default is 3.}
}
\description{
Common parameters used across codigo functions
}
\keyword{internal}
