% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eleven_map_to_one_ten}
\alias{eleven_map_to_one_ten}
\title{ICD 11 code or categories mapped to one ICD 10 category}
\format{
A data.frame with 17908 rows and 7 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{icd11_linearization_uri} \tab ICD 11 linearization URI \cr
   \emph{icd11_code} \tab ICD 11 code \cr
   \emph{icd11_chapter} \tab ICD 11 chapter \cr
   \emph{icd11_title} \tab ICD 11 title \cr
   \emph{icd10_code} \tab ICD 10 code (for class chapter) or block of codes (for class block) \cr
   \emph{icd10_chapter} \tab ICD 10 chapter \cr
   \emph{icd10_title} \tab ICD 10 class or kind: either category or code \cr
}
}
\source{
https://icdcdn.who.int/static/releasefiles/2024-01/mapping.zip
}
\usage{
eleven_map_to_one_ten
}
\description{
ICD 11 code or categories mapped to one ICD 10 category
}
\examples{
eleven_map_to_one_ten

}
\keyword{datasets}
