% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd11_linearization_mms}
\alias{icd11_linearization_mms}
\title{ICD-11 Mortality and Morbidity Linearization Outputs}
\format{
A data frame with 21 columns and 36753 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{Foundation URI} \tab Foundation URI \cr
   \emph{Linearization (release) URI} \tab Linearization release URI \cr
   \emph{Code} \tab ICD-11 Code \cr
   \emph{BlockId} \tab Block identifier \cr
   \emph{Title} \tab Title \cr
   \emph{ClassKind} \tab Class \cr
   \emph{DepthInKind} \tab Depth \cr
   \emph{IsResidual} \tab Is residual? \cr
   \emph{PrimaryLocation} \tab Primary location \cr
   \emph{ChapterNo} \tab Chapter number \cr
   \emph{BrowserLink} \tab Browser link \cr
   \emph{iCatLink} \tab Category link \cr
   \emph{isLeaf} \tab Is leaf? \cr
   \emph{noOfNonResidualChildren} \tab Number of non-residual children \cr
   \emph{Primary tabulation} \tab Primary tabulation \cr
   \emph{Version} \tab Date and time linearization was produced \cr
}
}
\source{
https://icd.who.int/dev11/Downloads/Download?fileName=LinearizationMiniOutput-MMS-en.zip
}
\usage{
icd11_linearization_mms
}
\description{
ICD-11 Mortality and Morbidity Linearization Outputs
}
\examples{
icd11_linearization_mms

}
\keyword{datasets}
