% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_10_get.R
\name{icd_10_get_releases}
\alias{icd_10_get_releases}
\alias{icd_10_get_chapters}
\alias{icd_10_get_release_by_category}
\alias{icd_10_get_info}
\title{Get available ICD-10 releases}
\usage{
icd_10_get_releases(
  api_version = c("v2", "v1"),
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_10_get_chapters(
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_10_get_release_by_category(
  category,
  api_version = c("v2", "v1"),
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_10_get_info(
  category,
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)
}
\arguments{
\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
For example, if you provide value v2, the API will respond in the format of
the version 2 of the API. Default is \code{v2}.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int. If you are using a locally deployed server or hosting
your own ICD API server, you should specify the URL of your instance here.}

\item{client}{The OAuth2 client produced through a call to
\code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{timeout}{The maximum time to wait for a response from the API server in
seconds. Default is 30 seconds.}

\item{max_tries}{The maximum number of times to retry the request in case of
failure. Default is 3.}

\item{release}{A string specifying the release version of ICD-10 to search
from. If not specified, defaults to the latest release version. See
the available versions with \code{icd_versions}.}

\item{language}{ICD-API is multi-lingual. By changing this header, you may
make the API respond in different languages. Languages will be available as
the translations of ICD-11 completes. The values are language codes such as
en, es, zh, etc. Depending on the \code{release_id} specified, the available
languages will vary. Default is English ("en"). Note that language support
for ICD-10 is limited to English (\code{en}).}

\item{verbose}{Logical. Should non-warning and non-error messages be
printed? Default is TRUE.}

\item{category}{ICD-10 category code or for blocks, the code range.}
}
\value{
A list with information on specified ICD 10 parameters
}
\description{
Get available ICD-10 releases
}
\examples{
\donttest{
  icd_10_get_releases()
  icd_10_get_chapters()
  icd_10_get_release_by_category(category = "A00")
  icd_10_get_release_by_category(category = "A00-A09")
  icd_10_get_info(category = "A00")
  icd_10_get_info(category = "A00-A09")
}

}
