% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_get_foundation.R
\name{icd_get_foundation}
\alias{icd_get_foundation}
\alias{icd_get_foundation_entity}
\title{Get information on various ICD-11 foundation entities}
\usage{
icd_get_foundation(
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_get_foundation_entity(
  id,
  release = NULL,
  include = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)
}
\arguments{
\item{release}{A string specifying the release version to search from. If not
specified, defaults to the latest available release version. See the
available versions for each ICD classification with \code{icd_versions}.}

\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
Default is \code{v2}.}

\item{language}{ICD-API is multi-lingual. Language codes such as "en", "es",
"zh", etc. Default is English ("en").}

\item{tabular}{Logical. Should output be structured into a tibble? Default
is TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be printed?
Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int.}

\item{client}{The OAuth2 client produced through a call to
\code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{timeout}{The maximum time to wait for a response from the API server in
seconds. Default is 30 seconds.}

\item{max_tries}{Maximum number of times to retry a request if it fails with
a transient error (e.g. 429, 500, 503). Default is 3.}

\item{id}{Unique numerical identifier for an entity.}

\item{include}{A string or a vector of strings for optional property values
to be included in the response. The property values that can be included
are \emph{"ancestor"}, \emph{"descendant"}, or \emph{"diagnosticCriteria"}. If not
specified, these properties are not included in the response.}
}
\value{
A list with information on specified ICD 11 Foundation and top level
entities.
}
\description{
Get information on various ICD-11 foundation entities
}
\examples{
\donttest{
  icd_get_foundation()
  icd_get_foundation_entity(id = "1435254666")     ## chapter
  icd_get_foundation_entity(id = "588616678")      ## block; depth 1
  icd_get_foundation_entity(id = "1465325727")     ## category; depth 1
}

}
