% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_map.R
\name{icd_map}
\alias{icd_map}
\title{Map ICD codes between versions}
\usage{
icd_map(from = c("icd10", "icd11"), to = c("one", "multiple"), code)
}
\arguments{
\item{from}{The ICD version to map from. Either \emph{icd10} or \emph{icd11}.
Default to \emph{icd10}.}

\item{to}{A character value of either \emph{multiple} or \emph{one} to map ICD-10
to ICD-11 accordingly. Only relevant if \code{from} is \emph{icd10}. Ignored when
\code{from} is \emph{icd11}. Default to \emph{one}.}

\item{code}{Corresponding ICD code in version specified in \code{from} to map
to the other version.}
}
\value{
A tibble of information for specified \code{code} mapped based on
\code{from}
}
\description{
Map ICD codes between versions
}
\examples{
icd_map(code = "A00")

}
