% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_versions}
\alias{icd_versions}
\title{Table of supported classifications and versions by the ICD API}
\format{
A data frame with 6 columns and 11 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{release_id} \tab Release identifier for specific classification version \cr
   \emph{foundation} \tab Logical. Does this version include the Foundation? \cr
   \emph{classification} \tab Classification name: ICD-11 or ICD-10 \cr
   \emph{linearization} \tab Either MMS or ICF \cr
   \emph{languages} \tab Languages in which classification, release, and linearization is available for \cr
}
}
\source{
https://icd.who.int/docs/icd-api/SupportedClassifications/
}
\usage{
icd_versions
}
\description{
Table of supported classifications and versions by the ICD API
}
\examples{
icd_versions

}
\keyword{datasets}
