% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ten_map_to_one_eleven}
\alias{ten_map_to_one_eleven}
\title{ICD 10 code or categories mapped to one ICD 11 category}
\format{
A data.frame with 12597 rows and 12 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{icd10_class_kind} \tab ICD 10 class or kind: either chapter, block, category, or modifiedcategory \cr
   \emph{icd10_depth} \tab ICD 10 depth level \cr
   \emph{icd10_code} \tab ICD 10 code (for class chapter) or block of codes (for class block) \cr
   \emph{icd10_chapter} \tab ICD 10 chapter \cr
   \emph{icd10_title} \tab ICD 10 class or kind: either category or code \cr
   \emph{icd11_class_kind} \tab ICD 11 class or kind: either block, category, or chapter \cr
   \emph{icd11_depth} \tab ICD 11 depth level \cr
   \emph{icd11_foundation_uri} \tab ICD 11 foundation URI \cr
   \emph{icd11_linearization_uri} \tab ICD 11 linearization URI \cr
   \emph{icd11_code} \tab ICD 11 code \cr
   \emph{icd11_chapter} \tab ICD 11 chapter \cr
   \emph{icd11_title} \tab ICD 11 title \cr
}
}
\source{
https://icdcdn.who.int/static/releasefiles/2024-01/mapping.zip
}
\usage{
ten_map_to_one_eleven
}
\description{
ICD 10 code or categories mapped to one ICD 11 category
}
\examples{
ten_map_to_one_eleven

}
\keyword{datasets}
