% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigo.R
\docType{package}
\name{codigo}
\alias{codigo-package}
\alias{codigo}
\title{Interface to the International Classification of Diseases (ICD) API}
\description{
The International Classification of Diseases (ICD) serves a broad range of
uses globally and provides critical knowledge on the extent, causes and
consequences of human disease and death worldwide via data that is reported
and coded with the ICD. ICD API allows programmatic access to the ICD. It is
an HTTP based REST API. This package provides functions that interface with
the ICD API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-ihtm.io/codigo/}
  \item \url{https://github.com/OxfordIHTM/codigo}
  \item Report bugs at \url{https://github.com/OxfordIHTM/codigo/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Anita Makori [copyright holder]
}

Other contributors:
\itemize{
  \item Sanjeev Pugazhendhi [thesis advisor, reviewer]
  \item Williams Udoh Ituen-Umanah [reviewer]
  \item Proochista Ariana [reviewer, copyright holder]
}

}
\keyword{internal}
