% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_authenticate.R
\name{icd_endpoint}
\alias{icd_endpoint}
\alias{icd_oauth_app}
\alias{icd_token}
\title{OAuth2 authentication to ICD API}
\usage{
icd_endpoint(
  authorize = NULL,
  access = "https://icdaccessmanagement.who.int/connect/token"
)

icd_oauth_app(
  appname = "icd_app",
  key = "6fc8a1e4-4da9-43a8-bd0c-c164c0cb0ebd_3c7e272e-2f4b-46de-b127-df7454e36be8",
  secret = "OGIly8mTatQ2ILuhBTbviIp2FofWiQR3fj4HaPiGseE="
)

icd_token(
  endpoint = icd_endpoint(),
  app = icd_oauth_app(),
  scope = "icdapi_access",
  as_header = TRUE,
  client_credentials = TRUE,
  ...
)
}
\arguments{
\item{authorize}{url to send client to for authorisation. Set to NULL if not
needed.}

\item{access}{url used to exchange unauthenticated for authenticated token.}

\item{appname}{Name of the application. This is not used for OAuth, but is
used to make it easier to identify different applications.}

\item{key}{Consumer key, also sometimes called the client ID.}

\item{secret}{Consumer secret, also sometimes called the client secret.
Despite its name, this does not necessarily need to be protected like a
password, i.e. the user still has to authenticate themselves and grant the
app permission to access resources on their behalf.}

\item{endpoint}{OAuth endpoint for the ICD API, created by \code{icd_endpoint()}}

\item{app}{An OAuth consumer application, created by \code{icd_oauth_app()} by
default.}

\item{scope}{ICD API scope to request. Default is "icdapi_access".}

\item{as_header}{If TRUE, the default, configures the token to add itself to
the bearer header of subsequent requests. If FALSE, configures the token to
add itself as a url parameter of subsequent requests.}

\item{client_credentials}{Default to FALSE. Set to TRUE to use Client
Credentials Grant instead of Authorization Code Grant. See
https://www.rfc-editor.org/rfc/rfc6749#section-4.4.}

\item{...}{Other parameters/arguments to be passed onto \code{httr::oauth2.0_token()}}
}
\value{
For \code{icd_token()}, a Token2.0 reference class (RC) object.
}
\description{
OAuth2 authentication to ICD API
}
\examples{
icd_endpoint()
icd_token()
icd_oauth_app()

}
