% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_autocode.R
\name{icd_autocode_foundation}
\alias{icd_autocode_foundation}
\alias{icd_autocode}
\title{Autocode against the foundation component or linearizations of the ICD-11}
\usage{
icd_autocode_foundation(
  q,
  subtree = NULL,
  release = NULL,
  threshold = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_autocode(
  q,
  linearization = c("mms", "icf"),
  subtree = NULL,
  release = NULL,
  threshold = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)
}
\arguments{
\item{q}{String. Text to be searched. Having the character \verb{\%} at the end
will be regarded as a wild card for that word.}

\item{subtree}{A string or vector of strings of URIs. If provided, the
search will be performed on the entities provided and their descendants.}

\item{release}{A string specifying the release version to search from. If not
specified, defaults to the latest available release version. See the
available versions for each ICD classification with \code{icd_versions}.}

\item{threshold}{A numeric value between 0 and 1 that indicates the
similarity between the input text and the matched term. \code{threshold} is the
minimum score to be included in the output. Default is NULL to use
default value specified by API.}

\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
Default is \code{v2}.}

\item{language}{ICD-API is multi-lingual. Language codes such as "en", "es",
"zh", etc. Default is English ("en").}

\item{tabular}{Logical. Should output be structured into a tibble? Default
is TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be printed?
Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int.}

\item{client}{The OAuth2 client produced through a call to
\code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{timeout}{The maximum time to wait for a response from the API server in
seconds. Default is 30 seconds.}

\item{max_tries}{Maximum number of times to retry a request if it fails with
a transient error (e.g. 429, 500, 503). Default is 3.}

\item{linearization}{A character value for which linearization to search.
Currently, the possible values are \emph{"mms"} and \emph{"icf"}. Default is \emph{"mms"}.}
}
\value{
A tibble of autocode results showing the search text, the matching
text, the code, URIs for the foundation and linearization entities, the
matching level, the matching score, and the matching type.
}
\description{
Autocode against the foundation component or linearizations of the ICD-11
}
\examples{
\donttest{
  icd_autocode_foundation("cholera")
  icd_autocode(q = "cholera")
  icd_autocode(q = "impairment", linearization = "icf")
}

}
