% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_autocode.R
\name{icd_autocode_foundation}
\alias{icd_autocode_foundation}
\alias{icd_autocode_mms}
\title{Perform autocoding}
\usage{
icd_autocode_foundation(
  q,
  subtree = NULL,
  release = NULL,
  threshold = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access"
)

icd_autocode_mms(
  q,
  subtree = NULL,
  release = NULL,
  threshold = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access"
)
}
\arguments{
\item{q}{String. Text to be searched. Having the character \verb{\%} at the end will
be regarded as a wild card for that word.}

\item{subtree}{A string or vector of strings of URIs. If provided, the
search will be performed on the entities provided and their descendants.}

\item{release}{A string specifying the release version of the Foundation to
search from. If not specified, defaults to the latest release version. See
the available versions with \code{icd_versions}.}

\item{threshold}{A numeric value between 0 and 1 that indicates the similarity
between the input text and the matched term. \code{threshold} is the
minimum score to be included in the output. Default is NULL to use
default value specified by API.}

\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
For example, if you provide value v2, the API will respond in the format of
the version 2 of the API. Default is \code{v2}.}

\item{language}{ICD-API is multi-lingual. By changing this header, you may
make the API respond in different languages. Languages will be available as
the translations of ICD-11 completes. The values are language codes such as
en, es, zh, etc. Depending on the \code{release_id} specified, the available
languages will vary. Default is English ("en").}

\item{tabular}{Logical. Should output be structured into a tibble? Default
to TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be
printed? Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int. If you are using a locally deployed server or hosting
your own ICD API server, you should specify the URL of your instance here.}

\item{client}{The OAuth2 client produced through a call to \code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}
}
\value{
An autocode
}
\description{
Perform autocoding
}
\examples{
icd_autocode_foundation("cholera")
icd_autocode_mms(release = "2024-01", q = "cholera")

}
