% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_check.R
\name{icd_check_release}
\alias{icd_check_release}
\alias{icd_check_language}
\title{Check that release date is specified correctly}
\usage{
icd_check_release(release)

icd_check_language(release = NULL, language)
}
\arguments{
\item{release}{A string specifying the release version of the ICD-11.}

\item{language}{language codes such as en, es, zh, etc.}
}
\value{
A message or an error if \code{release} provided is recognised or not. A
message or a warning if \code{language} requested is available or not for
\code{release} provided.
}
\description{
Check that release date is specified correctly
}
\examples{
icd_check_release("2024-01")
try(icd_check_release("2025-01"))
icd_check_language("2024-01", "ar")
icd_check_language("2024-01", "rr")
try(icd_check_language("2025-01", "ar"))

}
