% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_check.R
\name{icd_check_release}
\alias{icd_check_release}
\alias{icd_check_language}
\title{Checks for specified parameters supplied to search, autocode, and get
functions}
\usage{
icd_check_release(release, icd = c("icd11", "icd10"), verbose = TRUE)

icd_check_language(
  release = NULL,
  language,
  icd = c("icd11", "icd10"),
  verbose = TRUE
)
}
\arguments{
\item{release}{A string specifying the release version of the ICD-11.}

\item{icd}{A character string of available ICD classifications. Currently,
this can be either "icd10" or "icd11". Default is "icd11".}

\item{verbose}{Logical. Should non-warning and non-error messages be
printed? Default is TRUE.}

\item{language}{language codes such as en, es, zh, etc.}
}
\value{
A message or an error if \code{release} provided is recognised or not. A
message or a warning if \code{language} requested is available or not for
\code{release} provided. For \code{icd_check_language()}, returns a character value
or a character vector of acceptable language codes.
}
\description{
Checks for specified parameters supplied to search, autocode, and get
functions
}
\examples{
icd_check_release("2024-01")
try(icd_check_release("2025-01"))
icd_check_language("2024-01", "ar")
icd_check_language("2024-01", "rr")
try(icd_check_language("2025-01", "ar"))

}
