% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_get.R
\name{icd_get_foundation}
\alias{icd_get_foundation}
\alias{icd_get_entity}
\alias{icd_get_info}
\alias{icd_get_chapter}
\title{Get information on various ICD-11 foundation and linearization entities}
\usage{
icd_get_foundation(
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_get_entity(
  id,
  release = NULL,
  include = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_get_info(
  linearization = c("mms", "icf"),
  api_version = c("v2", "v1"),
  language = "en",
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)

icd_get_chapter(
  linearization = c("mms", "icf"),
  release = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access",
  timeout = 30,
  max_tries = 3
)
}
\arguments{
\item{release}{A string specifying the release version of the Foundation to
search from. If not specified, defaults to the latest release version. See
the available versions with \code{icd_versions}.}

\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
For example, if you provide value v2, the API will respond in the format of
the version 2 of the API. Default is \code{v2}.}

\item{language}{ICD-API is multi-lingual. By changing this header, you may
make the API respond in different languages. Languages will be available as
the translations of ICD-11 completes. The values are language codes such as
en, es, zh, etc. Depending on the \code{release_id} specified, the available
languages will vary. Default is English ("en").}

\item{tabular}{Logical. Should output be structured into a tibble? Default
to TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be
printed? Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int. If you are using a locally deployed server or hosting
your own ICD API server, you should specify the URL of your instance here.}

\item{client}{The OAuth2 client produced through a call to
\code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{timeout}{The number of seconds to wait for a response from the API.
Default is 30 seconds.}

\item{max_tries}{Maximum number of times to retry a request if it fails with a
transient error (e.g. 429, 500, 503). Default is 3.}

\item{id}{Unique numerical identifier for an entity.}

\item{include}{A string or a vector of strings for optional property values
to be included in the response. The property values that can be included
are \emph{"ancestor"}, \emph{"descendant"}, or \emph{"diagnosticCriteria"}. If not
specified, these properties are not included in the response.}

\item{linearization}{A character value for which linearization to search.
Currently, the possible values for this are \emph{"mms"} and \emph{"icf"}. Default
is \emph{"mms"}.}
}
\value{
A list with information on specified ICD 11 Foundation and top level
entities.
}
\description{
Get information on various ICD-11 foundation and linearization entities
}
\examples{
icd_get_foundation()
icd_get_entity(id = "1435254666")     ## chapter
icd_get_entity(id = "588616678")      ## block; depth 1
icd_get_entity(id = "1465325727")     ## category; depth 1
icd_get_info()
icd_get_info("icf")
icd_get_chapter()
icd_get_chapter("icf")

}
