% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_search.R
\name{icd_search_foundation}
\alias{icd_search_foundation}
\alias{icd_search}
\title{Search the foundation component of the ICD-11}
\usage{
icd_search_foundation(
  q,
  subtree = NULL,
  chapter = NULL,
  flexisearch = FALSE,
  flat = TRUE,
  properties = NULL,
  release = NULL,
  highlight = FALSE,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access"
)

icd_search(
  q,
  linearization = "mms",
  subtree = NULL,
  use_foundation = FALSE,
  keyword = FALSE,
  chapter = NULL,
  flexisearch = FALSE,
  flat = TRUE,
  release = NULL,
  highlight = FALSE,
  medical_mode = TRUE,
  properties = NULL,
  api_version = c("v2", "v1"),
  language = "en",
  tabular = TRUE,
  verbose = TRUE,
  base_url = "https://id.who.int",
  client = icd_oauth_client(),
  scope = "icdapi_access"
)
}
\arguments{
\item{q}{String. Text to be searched. Having the character \verb{\%} at the end will
be regarded as a wild card for that word.}

\item{subtree}{A string or vector of strings of URIs. If provided, the
search will be performed on the entities provided and their descendants.}

\item{chapter}{A string or vector of strings of chapter codes
eg: c("01", "02") When provided, the search will be performed only on
these chapters.}

\item{flexisearch}{Logical. Default is FALSE. Changes the search mode to
flexible search (for more details, see section on flexisearch).}

\item{flat}{Logical. Default is FALSE. If set to true the search result
entities are provided in a nested data structure representing the
ICD-11 hierarchy. Otherwise they are listed as flat list of matches.}

\item{properties}{A string or a vector of strings for the properties to be
searched. When set to NULL (default), search is performed on a default
set of properties depending on whether foundation or linearization (see
section on \code{properties} for more details). Should be set/specified if
\code{medical_mode = FALSE}.}

\item{release}{A string specifying the release version of the Foundation to
search from. If not specified, defaults to the latest release version. See
the available versions with \code{icd_versions}.}

\item{highlight}{Logical. Default is FALSE. If set to FALSE the search result
highlighting is turned off and the results don't contain special tags for
highlighting where the results are found within the text.}

\item{api_version}{Version of the API. Possible values are \code{v1} or \code{v2}.
For example, if you provide value v2, the API will respond in the format of
the version 2 of the API. Default is \code{v2}.}

\item{language}{Language code for language in which search results are to be
shown. Default is English ("en"). See section on \code{language} for more
details.}

\item{tabular}{Logical. Should output be structured into a tibble? Default
to TRUE.}

\item{verbose}{Logical. Should non-warning and non-error messages be
printed? Default is TRUE.}

\item{base_url}{The base URL of the API. Default uses the WHO API server at
https://id.who.int. If you are using a locally deployed server or hosting
your own ICD API server, you should specify the URL of your instance here.}

\item{client}{The OAuth2 client produced through a call to \code{icd_oauth_client()}.}

\item{scope}{Scopes to be requested from the resource owner. Default is
\emph{"icdapi_access"} as specified in the ICD API documentation.}

\item{linearization}{A character value for which linearization to search.
Currently, the only possible value for this is \emph{"mms"}. See section on
\code{linearization} for more details.}

\item{use_foundation}{Logical. Default is FALSE. Should subtree filter
use foundation descendants (TRUE) or linearization descendants (FALSE)?}

\item{keyword}{Logical. Default is FALSE. Should search results show keyword
list? See section on \code{keyword} for more details.}

\item{medical_mode}{Logical. Default is FALSE. Should medical coding
mode be turned on?}
}
\value{
A tibble of search results.
}
\description{
Search the foundation component of the ICD-11
}
\section{On \code{flexisearch}}{
 In the regular search mode (\code{flexisearch = FALSE}),
the function will only give you results that contain all of the words that
you've used in your search. It accepts different variants or synonyms of
the words but essentially it searches for a result that contains all
components of your search. Whereas in flexible search mode, the results do
not have to contain all of the words that are typed. It would still try to
find the best matching phrase but there may be words in your search that
are not matched at all. It is recommended to use flexible search only when
regular search does not provide a result.
}

\section{On \code{keyword}}{
 If set to true, the search result will also include
keyword list. If the last word provided is incomplete, keyword list
includes all words that start with the incomplete word
(word completion mode). If the last word is complete, the keyword list
will provide suggested additional words that could be added to the search
query (word suggestion mode).
}

\section{On \code{medical_mode}}{
 When searching the classification for
medical coding purposes, this should be set to TRUE (default). In this
mode, the search gives results only from the entities that have a code.
The system will search all index terms of an entity. i.e. \emph{titles},
\emph{synonyms}, \emph{fully specified term}, all terms of other entities that are in
the foundation are aggregated into this entity. By default, \emph{chapters 26},
\emph{V}, and \emph{X} are not included in the search results. If \code{medical_mode} is
set to FALSE, then the \code{properties} argument will need to be specified
(see next section).
}

\section{On \code{properties}}{
 In the Foundation search, by default the function
searches  \emph{"Title"}, \emph{"Synonyms"}, and \emph{"FullySpecifiedName"}. The valid
values that could be used for properties are: \emph{"Title"}, \emph{"Synonym"},
\emph{"NarrowerTerm"}, \emph{"FullySpecifiedName"}, \emph{"Definition"}, and
\emph{"Exclusion"}. In the MMS search, this argument is only used when
\code{medical_mode = FALSE}. The valid values that could be used are:
\emph{"Title"}, \emph{"FullySpecifiedName"}, \emph{"Definition"}, \emph{"Exclusion"}, and
\emph{"IndexTerm"}. If \emph{"IndexTerm"} is used, the search will be performed on
all \emph{"Titles"}, \emph{"Synonyms"}, and \emph{"FullySpecifiedNames"} including the
ones that are under shoreline (i.e Entities in the foundation but not in
MMS). In such cases the results will be shown based on where the match is
aggregated into in MMS.
}

\section{On \code{language}}{
 ICD-API is multi-lingual. By setting the language,
you may make the API respond in different languages. Languages will be
available as the translations of ICD-11 completes. The values are language
codes such as en, es, zh, etc. Depending on the \code{release} specified, the
available languages will vary.
}

\examples{
icd_search_foundation("colorectal cancer")
icd_search_mms("colorectal cancer")

}
