% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_structure.R
\name{icd_structure_foundation}
\alias{icd_structure_foundation}
\alias{icd_structure_search}
\title{Structure ICD list and ICD search outputs}
\usage{
icd_structure_foundation(icd_list)

icd_structure_search(icd_search)
}
\arguments{
\item{icd_list}{An \code{icd_list} object produced by a call to any of the
\code{icd_get} functions}

\item{icd_search}{An \code{icd_search} object produced by a call to any of the
\code{icd_search} functions}
}
\value{
A tibble of structured ICD outputs.
}
\description{
Structure ICD list and ICD search outputs
}
\examples{
icd_list <- icd_get_foundation(tabular = FALSE)
icd_tbl_foundation <- icd_structure_foundation(icd_list)
icd_structure_search(icd_search_foundation("cholera", tabular = FALSE))

}
