% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_structure.R
\name{icd_structure_foundation}
\alias{icd_structure_foundation}
\title{Structure ICD list outputs}
\usage{
icd_structure_foundation(icd_list)
}
\arguments{
\item{icd_list}{An \code{icd_list} object produced by a call to any of the \code{icd_get}
functions}
}
\value{
An object of class \code{icd_tbl} which is basically a \code{data.frame} format
of ICD outputs
}
\description{
Structure ICD list outputs
}
\examples{
icd_list <- icd_get_foundation()
icd_tbl_foundation <- icd_structure_foundation(icd_list)

}
