% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icd_get_releases}
\alias{icd_get_releases}
\title{Utility functions for retrieving ICD parameters and definitions}
\usage{
icd_get_releases(icd = c("icd11", "icd10"))
}
\arguments{
\item{icd}{A character string of available ICD classifications. Currently,
this can be either "icd10" or "icd11". Default is "icd11". If NULL, all ICD
classifications are used (currently both "icd10" and "icd11").}
}
\value{
A character value or a vector of character values for release
identifiers
}
\description{
Utility functions for retrieving ICD parameters and definitions
}
\examples{
icd_get_releases()

}
