% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_versions}
\alias{icd_versions}
\title{Table of supported classifications and versions by the ICD API}
\format{
A data frame with 6 columns and 11 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{release_id} \tab Release identifier for specific classification version \cr
   \emph{foundation} \tab Logical. Does this version include the Foundation? \cr
   \emph{mms} \tab Languages in which version and release is available for \cr
   \emph{icf} \tab Languages in which version and release is available for \cr
}


For logical fields that specify NA, this indicates that for that particular
version and release, the particular field was not applicable or relevant.
}
\source{
https://icd.who.int/docs/icd-api/SupportedClassifications/
}
\usage{
icd_versions
}
\description{
Table of supported classifications and versions by the ICD API
}
\examples{
icd_versions

}
\keyword{datasets}
